% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterpick.R
\name{counterpick}
\alias{counterpick}
\title{counterpick function}
\arguments{
\item{dat}{data}

\item{friendly}{vector of friendly champions}

\item{opponent}{vector of enemy champions}

\item{numPicks}{number of picks- default 1}
}
\value{
the champion on friendly team with highest win rate
}
\description{
counterpick function
}

