% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterpick_group.R
\name{counterpick_group}
\alias{counterpick_group}
\title{counterpick_group function, except that when filtering games, instead of specific champions, now it's a group of champions. The purpose is to increase sample size.}
\arguments{
\item{data, }{vector of friendly champions, vector of enemy champions, and number of picks- default 1}
}
\value{
the champion on friendly team with highest win rate
}
\description{
counterpick_group function, except that when filtering games, instead of specific champions, now it's a group of champions. The purpose is to increase sample size.
}

